refactor: use fields mapping system
