+++ libs/feature/search/src/lib/elasticsearch/index.ts
@@ -1,3 +1,3 @@
-export * from './elasticsearch.mapper'
+export * from './mapper'
 export * from './elasticsearch.service'
 export * from './fixtures'
