+++ libs/feature/search/src/lib/elasticsearch/mapper/index.ts
@@ -1,0 +1,2 @@
+export * from './elasticsearch.mapper'
+export * from './elasticsearch.field.mapper'
